refactor(app): Update tip length cal videos
These videos don't show the rest of the deck slots in the background,
and therefore are safe to use no matter where the calibration block is.
This is required for health check because the calibration block is in a
different place.
